refactor: restructure Backups module to views-based architecture
  Move all Backups-related code from src/app/backups to src/views/Backups following the new
  project structure:

  - Move BackupsView and all components to src/views/Backups/components/
  - Move custom hooks to src/views/Backups/hooks/
  - Move services to src/views/Backups/services/
  - Move utils to src/views/Backups/utils/
  - Update all import paths across the codebase
  - Add index.ts for proper exports
