+++ src/views/Backups/BackupsView.tsx
@@ -1,27 +1,27 @@
 import { useState } from 'react';
-import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
-import BreadcrumbsBackupsView from '../../../shared/components/Breadcrumbs/Containers/BreadcrumbsBackupsView';
-import { useAppDispatch, useAppSelector } from '../../../store/hooks';
+import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
+import BreadcrumbsBackupsView from 'app/shared/components/Breadcrumbs/Containers/BreadcrumbsBackupsView';
+import { useAppDispatch, useAppSelector } from 'app/store/hooks';
 import { Helmet } from 'react-helmet-async';
-import DeleteBackupDialog from '../../../drive/components/DeleteBackupDialog/DeleteBackupDialog';
-import WarningMessageWrapper from '../../../drive/components/WarningMessage/WarningMessageWrapper';
-import BackupsAsFoldersList from '../../components/BackupsAsFoldersList/BackupsAsFoldersList';
-import DeviceList from '../../components/DeviceList/DeviceList';
-import FileViewerWrapper from '../../../drive/components/FileViewer/FileViewerWrapper';
-import newStorageService from '../../../drive/services/new-storage.service';
-import { deleteFile } from '../../../drive/services/file.service';
-import { deleteItemsThunk } from '../../../store/slices/storage/storage.thunks/deleteItemsThunk';
-import { DriveItemData } from '../../../drive/types';
+import DeleteBackupDialog from 'app/drive/components/DeleteBackupDialog/DeleteBackupDialog';
+import WarningMessageWrapper from 'app/drive/components/WarningMessage/WarningMessageWrapper';
+import BackupsAsFoldersList from './components/BackupsAsFoldersList';
+import DeviceList from './components/DeviceList';
+import FileViewerWrapper from 'app/drive/components/FileViewer/FileViewerWrapper';
+import newStorageService from 'app/drive/services/new-storage.service';
+import { deleteFile } from 'app/drive/services/file.service';
+import { deleteItemsThunk } from 'app/store/slices/storage/storage.thunks/deleteItemsThunk';
+import { DriveItemData } from 'app/drive/types';
 import { DriveFolderData } from '@internxt/sdk/dist/drive/storage/types';
-import { contextMenuSelectedBackupItems } from '../../../drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
-import { useBackupListActions } from 'app/backups/hooks/useBackupListActions';
-import { useBackupDeviceActions } from 'app/backups/hooks/useBackupDeviceActions';
-import { useBackupsPagination } from 'app/backups/hooks/useBackupsPagination';
+import { contextMenuSelectedBackupItems } from 'app/drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
+import { useBackupListActions } from './hooks/useBackupListActions';
+import { useBackupDeviceActions } from './hooks/useBackupDeviceActions';
+import { useBackupsPagination } from './hooks/useBackupsPagination';
 import errorService from 'app/core/services/error.service';
 import notificationsService, { ToastType } from 'app/notifications/services/notifications.service';
 import { Dialog, MenuItemType } from '@internxt/ui';
-import { DownloadManager } from '../../../network/DownloadManager';
-import workspacesSelectors from '../../../store/slices/workspaces/workspaces.selectors';
+import { DownloadManager } from 'app/network/DownloadManager';
+import workspacesSelectors from 'app/store/slices/workspaces/workspaces.selectors';
 
 export default function BackupsView(): JSX.Element {
   const { translate } = useTranslationContext();
