+++ src/views/Backups/hooks/useBackupsPagination.ts
@@ -1,10 +1,10 @@
 import { useEffect, useState, useCallback } from 'react';
 import _ from 'lodash';
 import { t } from 'i18next';
-import newStorageService from '../../drive/services/new-storage.service';
-import { DriveItemData } from '../../drive/types';
-import errorService from '../../core/services/error.service';
-import notificationsService, { ToastType } from '../../notifications/services/notifications.service';
+import newStorageService from '../../../app/drive/services/new-storage.service';
+import { DriveItemData } from '../../../app/drive/types';
+import errorService from '../../../app/core/services/error.service';
+import notificationsService, { ToastType } from '../../../app/notifications/services/notifications.service';
 
 const DEFAULT_LIMIT = 50;
