+++ src/views/Backups/index.ts
@@ -1,0 +1,5 @@
+export { default } from './BackupsView';
+export { default as BackupsAsFoldersList } from './components/BackupsAsFoldersList';
+export { default as BackupDropdownActions } from './components/BackupDropdownActions';
+export { default as DeviceDropdownActions } from './components/DeviceDropdownActions';
+export { default as DeviceList } from './components/DeviceList';
