+++ src/views/Backups/services/backups.service.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it, Mock, vi } from 'vitest';
 import backupsService from './backups.service';
-import { SdkFactory } from '../../core/factory/sdk';
+import { SdkFactory } from '../../../app/core/factory/sdk';
 
-vi.mock('../../core/factory/sdk', () => ({
+vi.mock('../../../app/core/factory/sdk', () => ({
   SdkFactory: {
     getNewApiInstance: vi.fn(),
   },
