+++ src/views/Backups/utils/mappers.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest';
 import { mapBackupFolder } from './mappers';
 import { aes } from '@internxt/lib';
-import type { DriveFolderData } from '../../drive/types';
+import type { DriveFolderData } from '../../../app/drive/types';
 import envService from 'app/core/services/env.service';
 
 vi.mock('@internxt/lib', () => ({
