+++ src/views/Backups/utils/mappers.ts
@@ -1,5 +1,5 @@
 import { aes } from '@internxt/lib';
-import { DriveFolderData } from '../../drive/types';
+import { DriveFolderData } from 'app/drive/types';
 import { DriveFolderData as DriveFolderDataSDK } from '@internxt/sdk/dist/drive/storage/types';
 import envService from 'app/core/services/env.service';
